requirements.txt
fastapi
uvicorn
httpx==0.21.0
bcrypt==3.2.0
pyjwt==2.3.0
sqlalchemy==1.4.23
pytest==7.4.1
pytest-asyncio==0.17.2
bcrypt
scikit-learn==1.2.2
emoji==2.8.0
pandas
numpy
fasttext==0.9.2
python-multipart